Added dummy data and installed mysql2 dependencies
Dummy data has been created and successfully implemented into the database, meaning that testing with this data can now be done. I have also installed the mysql2 dependency as this seems to be the best way to use the mysql database in nextjs. The install for this is done through npm install mysql2 when you cd into eduquest.